:bug: jwt 롤백 진행
