+++ src/main/resources/config/application-prod.properties
@@ -41,4 +41,4 @@ swagger.api-key-name=Authorization
 swagger.api-pass-as=header
 swagger.authorization-scope=global
 swagger.autorization-description=accessEverything
-swagger.white-list=/v2/api-docs,/swagger-resources,/swagger-resources/**,/configuration/ui,/configuration/security,/swagger-ui.html,/webjars/**,/v3/api-docs/**,/swagger-ui/**
+swagger.white-list=/v2/api-docs,/swagger-resources,/swagger-resources/**,/configuration/ui,/configuration/security,/swagger-ui.html,/webjars/**,/v3/api-docs/**,/swagger-ui/**,/swagger
